// src/thor/model/ChatMessage.ts
export type ChatMessage  = {
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    sessionId?: string;
    /**
     * if this chat message is conencted to a service (ws)
     * @type {boolean}
     * @memberof ChatMessage
     */
    connected?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    json?: string;
    /**
     * type of the message source
     * @type {string}
     * @memberof ChatMessage
     */
    sourceType?: ChatMessageSourceTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    sourceOwner?: string;
    /**
     * 
     * @type {string}
     * @memberof ChatMessage
     */
    role?: ChatMessageRoleEnum;
    /**
     * the message to send to the chat service
     * @type {string}
     * @memberof ChatMessage
     */
    content?: string;
    /**
     * Unique identifier for object in the system
     * @type {string}
     * @memberof ChatMessage
     */
    id?: string;
    /**
     * UUID of owner of the object in the system
     * @type {string}
     * @memberof ChatMessage
     */
    ownerId?: string;
    /**
     * Date of object creation
     * @type {Date}
     * @memberof ChatMessage
     */
    createdDate?: Date;
    /**
     * Data, including hash of the key(s) used to encrypt this record.
     * @type {string}
     * @memberof ChatMessage
     */
    keyHash?: string;
    /**
     * Last user to access object
     * @type {string}
     * @memberof ChatMessage
     */
    lastAccessedById?: string;
    /**
     * Timestamp of last access of object
     * @type {Date}
     * @memberof ChatMessage
     */
    lastAccessedDate?: Date;
    /**
     * Unique identifier for user who last modifed the object in the system
     * @type {string}
     * @memberof ChatMessage
     */
    lastModifiedById?: string;
    /**
     * Date of last object modification
     * @type {Date}
     * @memberof ChatMessage
     */
    lastModifiedDate?: Date;
}
// ...
export enum ChatMessageSourceTypeEnum {
    API = 'api',
    SERVERLOG = 'server_log',
    P2P = 'p2p'
}
/**
* @export
* @enum {string}
*/
export enum ChatMessageRoleEnum {
    USER = 'user',
    ASSISTANT = 'assistant'
}